frontend is working

+++ web/src/components/UserPreferences/PreferenceArray.jsx
@@ -19,68 +19,36 @@ const PreferenceArray = () => {
         console.log("clicked delete on: ", index)
     }
 
+    if (userPreferences.thisUsersPreferences.length > 0 ) {
     return (
             <div>
-                { preferences && preferences.map((preference, index) => {
-                    console.log(preference);
+                {preferences && preferences.map((preference, index) => {
                     return <Grid key={`${index}consArray`} container spacing={2} border={1}>
-                            <Grid xs={12} md={3}>
-                                <p>  {preference.pref} </p>
-                            </Grid>
-                            <Grid xs={12} md={3} borderLeft={1} borderRight={1}>
-                                <p>  {preference.choice} </p>
-                            </Grid>
-                            <Grid xs={12} md={3}>
-                                <Button variant={"contained"} onClick={() => deletePref(index)} > Delete Preference  </Button>
-                            </Grid>
-                            <Grid xs={12} md={3} borderLeft={1} borderRight={1}>
-                                <Button variant={"contained"} onClick={() => editPref(index)} > Edit Preference  </Button>
-                            </Grid>
+                        <Grid xs={12} md={3}>
+                            <p>  { preference.pref.text } </p>
                         </Grid>
+                        <Grid xs={12} md={3} borderLeft={1} borderRight={1}>
+                            <p>  { preference.choice.text } </p>
+                        </Grid>
+                        <Grid xs={12} md={3}>
+                            <Button variant={"contained"} onClick={() => deletePref(index)}> Delete Preference </Button>
+                        </Grid>
+                        <Grid xs={12} md={3} borderLeft={1} borderRight={1}>
+                            <Button variant={"contained"} onClick={() => editPref(index)}> Edit Preference </Button>
+                        </Grid>
+                    </Grid>
                 })}
+
             </div>
-        );
-    /*
-    if (preferences) {
 
-        preferences.forEach((preference, index) => {
-             constructedArray.push(
-
-            <Grid key={`${index}consArray`} container spacing={2} border={1}>
-                <Grid xs={12} md={3}>
-                   <p>  {preference.key} </p>
-                 </Grid>
-                <Grid xs={12} md={3} borderLeft={1} borderRight={1}>
-                  <p>  {preference.value} </p>
-                </Grid>
-                <Grid xs={12} md={3}>
-                    <Button variant={"contained"} onClick={() => deletePref(index)} > Delete Preference  </Button>
-                </Grid>
-                <Grid xs={12} md={3} borderLeft={1} borderRight={1}>
-                    <Button variant={"contained"} onClick={() => editPref(index)} > Edit Preference  </Button>
-                </Grid>
-            </Grid>
-             )
-        })
+    )
     } else {
-        constructedArray.push(
-            <Grid key={`GuarderdArray`} container spacing={2} border={1}>
-                <Grid xs={12} md={3}>
-                    <p>  No preferences yet </p>
-                </Grid>
-                <Grid xs={12} md={3} borderLeft={1} borderRight={1}>
-                    <p>  Add preference below </p>
-                </Grid>
-            </Grid>
-
-
+        return (
+        <div>
+        <p> No preferences choosen </p>
+        </div>
         )
     }
-
-    return (
-        constructedArray
-    )
-     */
 
 }
